+++ ketos/tests/audio/test_waveform.py
@@ -77,7 +77,7 @@ def test_from_wav_zero_pad(sine_wave_file, sine_wave):
         # Verify some things about the warning
         assert len(w) == 1
         assert issubclass(w[-1].category, RuntimeWarning)
-        assert "Waveform padded with zeros to achieve desired length" in str(w[-1].message)
+        assert "Waveform padded with zeros to achieve the required length to compute the stft. 0 samples were padded on the left and 132300 samples were padded on the right" in str(w[-1].message)
         # Verify some things about the waveform
         sig = sine_wave[1][2*44100:] #the last 1 second of the sine wave
         sig = np.concatenate([sig,np.zeros(3*44100)]) #append 3 seconds of zeros
@@ -97,7 +97,7 @@ def test_from_wav_negative_offset(sine_wave_file, sine_wave):
         # Verify some things about the warning
         assert len(w) == 1
         assert issubclass(w[-1].category, RuntimeWarning)
-        assert "Waveform padded with zeros to achieve desired length" in str(w[-1].message)
+        assert "Waveform padded with zeros to achieve the required length to compute the stft. 88200 samples were padded on the left and 0 samples were padded on the right" in str(w[-1].message)
         # Verify some things about the waveform
         sig = sine_wave[1][:2*44100] #first 2 seconds of the sine wave
         sig = np.concatenate([np.zeros(2*44100),sig]) #append 2 seconds of zeros
@@ -117,7 +117,7 @@ def test_from_wav_negative_offset_pad_with_reflection(sine_wave_file, sine_wave)
         # Verify some things about the warning
         assert len(w) == 1
         assert issubclass(w[-1].category, RuntimeWarning)
-        assert "Waveform padded with its own reflection to achieve desired length" in str(w[-1].message)
+        assert "Waveform padded with its own reflection to achieve required length to compute the stft. 44100 samples were padded on the left and 0 samples were padded on the right" in str(w[-1].message)
         # Verify some things about the waveform
         sig = sine_wave[1][:2*44100] #first 2 seconds of the sine wave
         sig = np.concatenate([sig[44100:0:-1],sig[:]]) #pre-pend reflection